Update to support high resolutions (hardcoded max of 3840x2160) and resolve rendering issues when in fullscreen.

+++ src/r_draw.cpp
@@ -54,7 +54,7 @@ R_EmitEdge
 void R_EmitEdge(mvertex_t* pv0, mvertex_t* pv1)
 {
     edge_t *edge, *pcheck;
-    int u_check;
+    int64_t u_check; // Changed from int to int64_t
     float u, u_step;
     vec3_t local, transformed;
     float* world;
