+++ src/render.h
@@ -57,8 +57,8 @@ typedef struct {
     //  for use in edge list
     float fvrectx, fvrecty;         // for floating-point compares
     float fvrectx_adj, fvrecty_adj; // left and top edges, for clamping
-    int vrect_x_adj_shift20;        // (vrect.x + 0.5 - epsilon) << 20
-    int vrectright_adj_shift20;     // (vrectright + 0.5 - epsilon) << 20
+    int64_t vrect_x_adj_shift20;        // (vrect.x + 0.5 - epsilon) << 20
+    int64_t vrectright_adj_shift20;     // (vrectright + 0.5 - epsilon) << 20
     float fvrectright_adj, fvrectbottom_adj;
     // right and bottom edges, for clamping
     float fvrectright;           // rightmost edge, for Alias clamping
